requirements.txt
appdirs==1.4.4
asgiref==3.8.1
astroid==2.3.3
Brotli==1.1.0
certifi==2020.4.5.1
cffi==1.14.3
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
cryptography==3.2
distlib==0.3.0
dj-database-url==2.2.0
Django==4.2.13
django-crispy-forms==1.14.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
filelock==3.0.12
gunicorn==20.1.0
h11==0.14.0
idna==3.7
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mpmath==1.1.0
onesignal-python-api==2.0.2
pipenv==2020.11.15
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.20
pylint==2.4.4
python-dateutil==2.9.0.post0
pytz==2020.1
requests==2.32.3
six==1.14.0
sqlparse==0.3.1
sympy==1.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
virtualenv==20.0.21
virtualenv-clone==0.5.4
Wave==0.0.2
whitenoise==6.0.0
wrapt==1.11.2
